requirements.txt
python-gitlab
PyGithub
matplotlib
seaborn
python-dotenv
rich
streamlit